aider: Include the templates directory in the xmllint command.
# Aider chat conversation:

USER: i appear to be getting the following error when running `warning: failed to load external entity "supplemental_oxygen_plan_clinician.opt"`

ASSISTANT: The error message you're seeing typically occurs when the file you're trying to access doesn't exist or isn't in the expected location. In this case, it seems like the `xmllint` command is unable to find the `supplemental_oxygen_plan_clinician.opt` file.

The issue might be that the `xmllint` command is being run from a different directory than where the `.opt` files are located. If the `.opt` files are in the `templates` directory, we need to include that in the command.

Here's the change:

```python
opt_validator.py
<<<<<<< ORIGINAL
        # Run the xmllint command with the specified xsd file
        command = f'xmllint --noout --schema {args.xsd} {opt_file}'
=======
        # Run the xmllint command with the specified xsd file
        # Include the templates directory in the command
        command = f'xmllint --noout --schema {args.xsd} templates/{opt_file}'
>>>>>>> UPDATED
```

This change ensures that the `xmllint` command looks in the `templates` directory for the `.opt` files.